Add files via upload
A number of useful scripts for the DBA in building and cloning systems in Oracle. Last used in Oracle 11gR2.

"re" in this context means reverse engineering. By running these ahead of time, you can spool the output and save yourself the trouble of  wading through exports or expdbs  to replicate data systematically.